requirements.txt
absl-py==1.2.0
appdirs==1.4.4
asgiref==3.5.2
astor==0.8.0
astunparse==1.6.3
attrs==22.2.0
audioread==2.1.8
backend==0.2.4.1
bootstrap4==0.1.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
chatterbot-corpus==1.2.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.10.0
decorator==4.4.0
distlib==0.3.6
Django==4.1.1
easyocr==1.6.2
filelock==3.9.0
flatbuffers==23.5.26
fontawesomefree==6.2.1
fonttools==4.37.3
gast==0.4.0
google-api-core==2.11.0
google-auth==2.18.1
google-auth-oauthlib==0.4.6
google-cloud-vision==3.4.1
google-pasta==0.2.0
googleapis-common-protos==1.59.0
grpcio==1.54.2
grpcio-status==1.54.2
gunicorn==20.1.0
h5py==3.7.0
idna==3.4
imageio==2.26.0
imutils==0.5.4
Jinja2==3.1.2
joblib==1.2.0
keras==2.9.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
lazy_loader==0.1
libclang==14.0.6
librosa==0.9.2
llvmlite==0.39.1
Markdown==3.4.1
MarkupSafe==2.1.1
mathparse==0.1.2
matplotlib==3.6.0
mediapipe==0.9.1.0
mpmath==1.3.0
networkx==3.0
ninja==1.11.1
nltk==3.8.1
numba==0.56.2
numpy==1.23.3
oauthlib==3.2.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.4.60
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
Pillow==9.2.0
Pint==0.22
platformdirs==2.6.2
pooch==1.6.0
proto-plus==1.22.2
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post4
pycparser==2.19
pymongo==3.13.0
pyparsing==2.4.2
pystache==0.6.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.7.5
pytz==2022.7
PyWavelets==1.4.1
pywin32==305
PyYAML==3.13
regex==2023.6.3
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.9
scikit-image==0.20.0
scikit-learn==1.1.2
scipy==1.9.1
shapely==2.0.1
six==1.12.0
SoundFile==0.10.2
split-image==2.0.1
SQLAlchemy==1.2.19
sqlparse==0.3.0
static==1.1.1
sympy==1.11.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.2
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.0.1
threadpoolctl==3.1.0
tifffile==2023.3.15
torch==2.0.0
torchvision==0.15.1
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.4
urllib3==1.26.12
virtualenv==20.17.1
Wave==0.0.2
Werkzeug==2.2.2
wrapt==1.14.1
